requirements.txt
requests~=2.32.4
pillow~=11.3.0
fastapi~=0.118.0